requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altgraph==0.17.4
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
edge-tts==7.0.2
etelemetry==0.3.1
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.0.0
frozenlist==1.7.0
future==1.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
multidict==6.6.3
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
pillow==11.3.0
propcache==0.3.2
prov==2.1.1
puremagic==1.30
pydot==4.0.1
pydub==0.25.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyMuPDF==1.26.3
PyMuPDFb==1.23.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.4
scipy==1.16.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
srt==3.5.3
tabulate==0.9.0
traits==7.0.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.0.1
yarl==1.20.1
